+++ source/Private/Test-FileContainsClassResource.ps1
@@ -2,14 +2,22 @@
     .SYNOPSIS
         Tests if a PowerShell file contains a DSC class resource.
 
+    .DESCRIPTION
+        This function parses a PowerShell file using the Abstract Syntax Tree (AST)
+        to determine if it contains any class-based DSC resources. It looks for 
+        class definitions that have a [DscResource] attribute, regardless of any
+        parameters the attribute may have. If any parsing errors are encountered,
+        an exception will be thrown with details about the error location.
+
     .PARAMETER FilePath
         The full path to the file to test.
 
     .EXAMPLE
-        Test-ContainsClassResource -ModulePath 'c:\mymodule\myclassmodule.psm1'
+        Test-FileContainsClassResource -FilePath 'c:\mymodule\myclassmodule.psm1'
 
-        This command will test myclassmodule for the presence of any class-based
-        DSC resources.
+        This command will test myclassmodule.psm1 for the presence of any class-based
+        DSC resources, including those with attributes like [DscResource()] or
+        [DscResource(RunAsCredential = 'Optional')].
 #>
 function Test-FileContainsClassResource
 {
@@ -22,13 +30,34 @@ function Test-FileContainsClassResource
         $FilePath
     )
 
-    $fileAst = [System.Management.Automation.Language.Parser]::ParseFile($FilePath, [ref]$null, [ref]$null)
-
-    foreach ($fileAttributeAst in $fileAst.FindAll( {$args[0] -is [System.Management.Automation.Language.AttributeAst]}, $false))
+    $tokens = $null
+    $parseErrors = $null
+    $fileAst = [System.Management.Automation.Language.Parser]::ParseFile($FilePath, [ref] $tokens, [ref] $parseErrors)
+    
+    # Check for parsing errors and throw exception for the first error
+    if ($parseErrors -and $parseErrors.Count -gt 0)
     {
-        if ($fileAttributeAst.Extent.Text -ieq '[DscResource()]')
+        $firstError = $parseErrors[0]
+        throw "Parse error in file '$FilePath': $($firstError.Message) at line $($firstError.Extent.StartLineNumber), column $($firstError.Extent.StartColumnNumber)"
+    }
+
+    # Look for class/type definitions that have a [DscResource(...)] attribute
+    $typeDefinitionAsts = $fileAst.FindAll({
+        param($ast)
+        $ast -is [System.Management.Automation.Language.TypeDefinitionAst]
+    }, $true)
+
+    foreach ($typeDefinitionAst in $typeDefinitionAsts)
+    {
+        foreach ($attributeAst in $typeDefinitionAst.Attributes)
         {
-            return $true
+            $attributeName = $attributeAst.TypeName.Name
+
+            if ($attributeName -eq 'DscResource')
+            {
+                # Return on first hit.
+                return $true
+            }
         }
     }